refactor: improving request validation and client methods.
- add a nil check to the Validate() method
- replace int64 with int in client methods and schemas
- replace bool with *bool in requests to allow null values
- update Patch() method (client) to handle different body types and perform validation
- improve null check in ListOptionsProvider
- add helper functions etc

+++ crowdin/storage.go
@@ -69,7 +69,7 @@ func (s *StorageService) List(ctx context.Context, opts *model.ListOptions) ([]*
 
 // Get returns a file in the storage by its identifier.
 // https://developer.crowdin.com/api/v2/#operation/api.storages.get
-func (s *StorageService) Get(ctx context.Context, id int64) (*model.Storage, *Response, error) {
+func (s *StorageService) Get(ctx context.Context, id int) (*model.Storage, *Response, error) {
 	res := new(model.StorageGetResponse)
 	resp, err := s.client.Get(ctx, fmt.Sprintf("/api/v2/storages/%d", id), nil, res)
 
@@ -78,6 +78,6 @@ func (s *StorageService) Get(ctx context.Context, id int64) (*model.Storage, *Re
 
 // Delete deletes a file from the storage by its identifier.
 // https://developer.crowdin.com/api/v2/#operation/api.storages.delete
-func (s *StorageService) Delete(ctx context.Context, id int64) (*Response, error) {
+func (s *StorageService) Delete(ctx context.Context, id int) (*Response, error) {
 	return s.client.Delete(ctx, fmt.Sprintf("/api/v2/storages/%d", id))
 }
